Fixes Issue #1081
The App crashed because the Adapter was trying to call a method on a
null object.

Where was the null object?
CheckBox was null because in the Layout file
view_key_adv_user_id_item.xml

There is no such widget with id = user_id_item_check_box

Fix :
I removed the two lines which were causing the bug, and now the screen
loads perfectly.