Make bright toolbar themeable
Also remove a superfluous divider in the decrypt files fragment and the
unused drawer_custom_header.xml.

This changes the color of the toolbar in the light theme as well,
because its brightness didn't play well with the system status bar of
Android, drowning it in whiteness. Now it is still very light, but the
system bar is still faintly visible above it.