Make import keys activity honour the theme
This changes the light theme slightly as well, it uses the actual
background instead of pure white and removes the thick superfluous
divider on top.
The divider at the bottom now has no margin, making the separation from
the FrameLayout on top more consistent, especially when there is
scrollable content.